+++ packages/react-components/src/index.ts
@@ -1,6 +1,7 @@
 export * from './useConnect';
 export * from './useConnection';
 export * from './useContractSelector';
+export * from './useGrpcClient';
 export * from './useWalletConnectorSelector';
 export * from './WithWalletConnector';
 export * from '@concordium/wallet-connectors';