+++ packages/react-components/src/useContractSelector.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { AccountAddress, CcdAmount, JsonRpcClient } from '@concordium/web-sdk';
+import { AccountAddress, CcdAmount, ConcordiumGRPCClient, JsonRpcClient } from '@concordium/web-sdk';
 import { errorString } from './error';
 
 /**
@@ -42,7 +42,7 @@ export interface Info {
     moduleRef: string;
 }
 
-async function refresh(rpc: JsonRpcClient, index: bigint) {
+async function refresh(rpc: ConcordiumGRPCClient | JsonRpcClient, index: bigint) {
     const info = await rpc.getInstanceInfo({ index, subindex: BigInt(0) });
     if (!info) {
         throw new Error(`contract ${index} not found`);
@@ -72,7 +72,7 @@ function parseContractIndex(input: string) {
     }
 }
 
-async function loadContract(rpc: JsonRpcClient, input: string) {
+async function loadContract(rpc: ConcordiumGRPCClient | JsonRpcClient, input: string) {
     const index = parseContractIndex(input);
     return refresh(rpc, index);
 }
@@ -105,7 +105,10 @@ export interface ContractSelector {
  * @param input The index of the contract to look up.
  * @return The resolved contract and related state.
  */
-export function useContractSelector(rpc: JsonRpcClient | undefined, input: string): ContractSelector {
+export function useContractSelector(
+    rpc: ConcordiumGRPCClient | JsonRpcClient | undefined,
+    input: string
+): ContractSelector {
     const [selected, setSelected] = useState<Info>();
     const [isLoading, setIsLoading] = useState(false);
     const [error, setError] = useState('');