+++ packages/wallet-connectors/src/BrowserWallet.ts
@@ -91,6 +91,27 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
     }
 
     /**
+     * Returns a gRPC client that is ready to perform requests against some Concordium Node connected to network/chain
+     * that the connected account lives on.
+     * The client implements version 2 of the Node API.
+     *
+     * This method is included because it's part of the Browser Wallet API.
+     * It should be used with care as it's hard to guarantee that it actually connects to the expected network.
+     * The recommended alternative is to have the application instantiate its own instance
+     * that is independent of any connection.
+     * See {@link Network.grpcOpts} for more details.
+     *
+     * Implementation detail: The method cannot be moved to {@link WalletConnector}
+     * as the Browser Wallet's RPC client doesn't work until a connection has been established.
+     *
+     * @return The Browser Wallet's internal gRPC client.
+     * @throws If the installed version of the Browser Wallet doesn't support the method.
+     */
+    getGrpcClient() {
+        return this.client.getGrpcClient();
+    }
+
+    /**
      * Deregister event handlers on the API client and notify the delegate.
      * As there's no way to actually disconnect the Browser Wallet, this is all that we can reasonably do.
      * The client object will remain in the browser's global state.