+++ packages/wallet-connectors/src/WalletConnection.ts
@@ -7,6 +7,7 @@ import {
     SchemaVersion,
     toBuffer,
 } from '@concordium/web-sdk';
+import { GrpcWebOptions } from '@protobuf-ts/grpcweb-transport';
 
 export type ModuleSchema = {
     type: 'ModuleSchema';
@@ -153,15 +154,25 @@ export interface WalletConnection {
      * Returns a JSON-RPC client that is ready to perform requests against some Concordium Node connected to network/chain
      * that the connected account lives on.
      *
-     * This method is included because it's part of the Browser Wallet's API.
+     * This method is included because it's part of the Browser Wallet API.
      * It should be used with care as it's hard to guarantee that it actually connects to the expected network.
-     * The application may easily instantiate its own client and use that instead for more control.
+     * As explained in {@link Network.jsonRpcUrl}, the application may easily instantiate its own client and use that instead.
      *
-     * Note that this method cannot be moved to {@link WalletConnector} as the Browser Wallet's RPC client doesn't work
-     * until a connection has been established.
+     * Implementation detail: The method cannot be moved to {@link WalletConnector}
+     * as the Browser Wallet's internal client doesn't work until a connection has been established.
      *
-     * @return Returns a JSON-RPC client for performing requests against a Concordium Node connected
-     * to the appropriate network.
+     * The client implements version 1 of the Node's API which is deprecated in favor of
+     * <code>ConcordiumGRPCClient</code> from {@link https://www.npmjs.com/package/@concordium/web-sdk @concordium/web-sdk}.
+     *
+     * The method {@link BrowserWalletConnector.getGrpcClient} exists for exposing the wallet's internal gRPC Web client for API version 2,
+     * but it's not recommended for use in most cases as there's no need for the client to be associated with a particular connection.
+     *
+     * Instead, instantiate the client manually or using a hook (for React) as described in {@link Network.grpcOpts}.
+     *
+     * @return A JSON-RPC client for performing requests against a Concordium Node connected to the appropriate network.
+     * @throws If the connection uses {@link Network.jsonRpcUrl} and that value is undefined.
+     * @throws If the connection is to the Browser Wallet and the installed version doesn't support the method.
+     * @deprecated Use <code>ConcordiumGRPCClient<code> instead.
      */
     getJsonRpcClient(): JsonRpcClient;
 
@@ -227,13 +238,49 @@ export interface Network {
     genesisHash: string;
 
     /**
-     * The URL of a <a href="https://github.com/Concordium/concordium-json-rpc">Concordium JSON-RPC proxy</a> instance
+     * The connection configuration for a gRPC Web endpoint for performing API (v2) queries against a
+     * Concordium Node instance connected to this network.
+     *
+     * The initialization is straightforward:
+     * <pre>
+     *   import { ConcordiumGRPCClient } from '@concordium/web-sdk';
+     *   import { GrpcWebFetchTransport } from '@protobuf-ts/grpcweb-transport';
+     *   ...
+     *   const client = new ConcordiumGRPCClient(new GrpcWebFetchTransport(network.grpcOpts!));
+     * </pre>
+     *
+     * For React projects, the hook <code>useGrpcClient</code> in
+     * {@link https://www.npmjs.com/package/@concordium/react-components @concordium/react-components}
+     * makes it very easy to obtain a client that's always connecting to the expected network.
+     */
+    grpcOpts: GrpcWebOptions | undefined;
+
+    /**
+     * The URL of a {@link https://github.com/Concordium/concordium-json-rpc Concordium JSON-RPC proxy} instance
      * for performing API (v1) queries against a Concordium Node instance connected to this network.
+     *
+     * The value is currently used only for {@link WalletConnectConnection WalletConnect connections}
+     * as {@link BrowserWalletConnector Browser Wallet connections} use the Browser Wallet's internal client.
+     *
+     * Setting the URL to the empty string disables the automatic initialization of the client returned by
+     * {@link WalletConnection.getJsonRpcClient} for WalletConnect connections.
+     * The concrete implementation {@link WalletConnectConnection.getJsonRpcClient} will then throw an exception
+     * as there is no client instance to return.
+     *
+     * There is no fundamental difference between using a client belonging to a connection compared to one that is initialized directly.
+     * Keeping it separate from connections gives the dApp more control and allows the client to be used even if no connections have been established.
+     *
+     * Initializing a separate client is straightforward:
+     * <pre>
+     *   import { HttpProvider, JsonRpcClient } from '@concordium/web-sdk';
+     *   ...
+     *   const client = new JsonRpcClient(new HttpProvider(network.jsonRpcUrl));
+     * </pre>
      */
     jsonRpcUrl: string;
 
     /**
-     * The base URL of a <a href="https://github.com/Concordium/concordium-scan">CCDScan</a> instance
+     * The base URL of a {@link https://github.com/Concordium/concordium-scan CCDScan} instance
      * connected to this network.
      * While CCDScan supports queries against its backend,
      * the main use of this URL is to construct links to the frontend.